Moves validateMarkdown to it's own file
